게시물 리스트에서 링크로 상세페이지 들어가기

+++ src/main/java/com/ll/exam/article/ArticleRepository.java
@@ -23,7 +23,7 @@ public class ArticleRepository {
         return datum;
     }
 
-    public ArticleDto findById(int id) {
+    public ArticleDto findById(long id) {
         for (ArticleDto articleDto : datum) {
             if(articleDto.getId() == id){
                 return articleDto;
